Moved extra camera logic to client

+++ Emu/include/CommonFunctions.h
@@ -26,7 +26,7 @@ namespace Engine
 		return ss.str();
 	}
 
-	inline static float Lerp(float a, float b, float f)
+	inline static double Lerp(double a, double b, double f)
 	{
 		return (a * (1.0f - f)) + (b * f);
 	}
